apps/api/pyproject.toml
[project]
name = "psma-api"
version = "0.0.0"
description = "PSMA backend API (FastAPI)"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.110",
  "httpx>=0.26",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.1",
]

[dependency-groups]
dev = [
  "jsonschema>=4.21",
  "pytest>=7.4",
  "ruff>=0.6",
]

[tool.ruff]
target-version = "py312"
exclude = [".venv", ".pytest_cache"]

[tool.ruff.lint]
# Police logging discipline: forbid print/pprint in application code.
select = ["T20"]

[tool.ruff.lint.per-file-ignores]
# CLI-like helper: printing is acceptable here.
"psma_api/export_openapi.py" = ["T201"]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["psma_api"]

[tool.pytest.ini_options]
testpaths = ["tests"]

